requirements.txt
# requirements file
# pip3 install -r requirements.txt
asn1crypto==1.5.1
base58==2.1.1
bcrypt==3.2.0
bech32==1.2.0
bit==0.7.2
bitstring==4.0.1
btc-hd-wallet==0.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coincurve==18.0.0
cryptography==3.4.8
ecdsa==0.18.0
gnupg==2.3.1
httplib2==0.20.2
idna==3.4
importlib-metadata==5.2.0
jeepney==0.8.0
keyring==23.5.0
libsecp256k1==0.1
mnemonic==0.20
numpy==1.24.0
opencv-python==4.6.0.66
psutil==5.9.3
pycparser==2.21
pylibscrypt==2.0.0
pyparsing==3.0.9
python-gnupg==0.5.0
qrcode==6.1
requests==2.28.1
SecretStorage==3.3.3
simplejson==3.18.0
six==1.16.0
sounddevice==0.4.5
urllib3==1.26.5
zipp==3.11.0
ripemd-hash==1.0.0
